Moved `AttributeSet` reference to opentelemetry root
