+++ opentelemetry-sdk/benches/metric.rs
@@ -2,7 +2,7 @@ use rand::Rng;
 use std::sync::{Arc, Weak};
 
 use criterion::{criterion_group, criterion_main, Bencher, Criterion};
-use opentelemetry::attributes::AttributeSet;
+use opentelemetry::AttributeSet;
 use opentelemetry::{
     metrics::{Counter, Histogram, MeterProvider as _, Result},
     Key, KeyValue,
