+++ opentelemetry-sdk/benches/metric_counter.rs
@@ -1,5 +1,5 @@
 use criterion::{criterion_group, criterion_main, Criterion};
-use opentelemetry::attributes::AttributeSet;
+use opentelemetry::AttributeSet;
 use opentelemetry::{
     metrics::{Counter, MeterProvider as _},
     KeyValue,
