+++ opentelemetry-sdk/src/metrics/internal/histogram.rs
@@ -2,7 +2,7 @@ use std::{collections::HashMap, sync::Mutex, time::SystemTime};
 
 use crate::metrics::data::HistogramDataPoint;
 use crate::metrics::data::{self, Aggregation, Temporality};
-use opentelemetry::{attributes::AttributeSet, global, metrics::MetricsError};
+use opentelemetry::{global, metrics::MetricsError, AttributeSet};
 
 use super::{
     aggregate::{is_under_cardinality_limit, STREAM_OVERFLOW_ATTRIBUTE_SET},
